Combine all rich-text formatted substrings when reading the string table.
Ensures all separately formatted substrings in a cell are concatenated together as the CT_RElt cell type specifies.

NOTE: In the case <t> and <r> elements are mixed, we read the <t> element
first, and then append any <r> elements. This is consistent with the spec
(CT_RElt) but is not consistent with how either Excel2010 or MacOSX
Preview handle xlsx strings:

Excel2010: <t>a</t><r><t>b</t></r> is read as ab, while
<r><t>a</t></r><t>b</t> is invalid.

MacOSX Preview: Ignores <r> elements if a <t> element is present.