Make xls and xlsx more similar w.r.t. stringTable
Another preparatory step towards de-duplicating code that operates on the cells

For now, at least, just give xlsWorkbook a dummy stringTable.

Add the workbook's stringTable as a parameter of inferType(), asCharSxp(), asStdString() for xls, for parallelism with xlsx.

Make stringTable the last argument of those functions. The vague motivation for this is that it feels favorable for perhaps eventually making stringTable an optional argument.