Use a template class to remove xls/xlsx duplication  (#683)
* Hello, World of Templates!

* Introduce SheetData

The jobs of the old Xls[x]WorkSheet class are going to be split up between class template Sheet<T> and Xls[x]SheetData.

* Use existing Xls[x]WorkBook classes

* Make insertShims() and advance_row() into template functions

* Make xls and xlsx more similar w.r.t. stringTable

Another preparatory step towards de-duplicating code that operates on the cells

For now, at least, just give xlsWorkbook a dummy stringTable.

Add the workbook's stringTable as a parameter of inferType(), asCharSxp(), asStdString() for xls, for parallelism with xlsx.

Make stringTable the last argument of those functions. The vague motivation for this is that it feels favorable for perhaps eventually making stringTable an optional argument.

* Real definition for XlsSheetData

* Real definition for XlsxSheetData

* Implement colNames() in Sheet<T>

* Remove more toy stuff from start

* Expose fully-featured col_names argument

* Implement colTypes(), expose col_types and guess_max arguments

* Implement readCols() in Sheet<T>

* Cut over to the new code

* Rename Xls[x]SheetData to Xls[x]CellSet

This better reflects the data managed by this class.

Also nice because "SheetData" comes up in the XML we work with on the xlsx side. So removes chance of confusion there.

* Rename file

* Skip the 1 test (out of 391) that is failing

* Go back to sheet_ belonging to the class

* Document why sheet_ must belong to XlsxCellSet

* No explicit destructor needed for XlsxCellSet

* Rename Sheet<T> to SheetView<T>

* Remove print statements

* Move definition closer to first use

* Function not used anywhere

* Use #pragma once

* Transfer ownership of the spinner to SheetView<T>