Fix compilation on Alpine Linux
and other systems that use musl.

The problem is that the internal musl header that defines
uid_t and gid_t only defines it conditionally, if a macro is set.
This macro is set if the header is included from unistd.h,
but not if included from another header.

readxl uses `#pragma once`, so the header will not be included
twice, so `uid_t` might not be defined, if unistd.h is not
included early on.

The same story happens with `struct timeval` and `sys/time.h`.