Use a cell's self-declared location, even if it has no content we will use
Fixes #648

The example file is presumably created by a 3rd party tool. Another one of those tool's that (mostly) does not declare each cell's location and we have to (mostly) infer from location within the sheetData.

Here's what the start of the (first) afflicted row looks like in the sheetData node of the XML:

<row>
  [1] <c r="B18043" s="52"/>
  [2] <c s="33"/>
  [3] <c s="28" t="s">\n  <v>19594</v>\n</c>
  [4] <c s="154" t="s">\n  <v>2318</v>\n</c>
  [5] <c s="101" t="s">\n  <v>19550</v>\n</c>

Previously we were not using the B18043 location information.
So the first content-ful cell in this row was interpreted as belonging to the 3rd column (C), but it's really in the 4th column (D).

Now we take advantage of any cell that declares its location, even if we ultimately won't load this cell.