Use a row's self-declared number
Fixes #671

The example file is presumably created by a 3rd party tool. Another one of those tool's that (mostly) does not declare each cell's location and we have to (mostly) infer from location within the sheetData.

But the XML does always provide an explicit row number in the row after a completely missing row. We just weren't using this information before.

More detailed looks at the XML are given in the issue, where I originally said I wouldn't fix this. But then I realized it was a pretty easy fix.