Add a test; tweak some tests
Add a test and test file based on a format where the confusion can come via currency, not colour.

Link to issues.

We generally incorporate `-xls` or `-xlsx` into test file names, so the format is obvious even if the file is open in Excel.

Added the actual implicit number formats in comments.